+++ BattleCityRemastered.vcxproj
@@ -174,23 +174,23 @@
     <ClInclude Include="headers\GameFailure.h" />
     <ClInclude Include="headers\GameStatistics.h" />
     <ClInclude Include="headers\GameSuccess.h" />
-    <ClInclude Include="headers\IGame.h" />
-    <ClInclude Include="headers\IDrawable.h" />
-    <ClInclude Include="headers\IMovable.h" />
-    <ClInclude Include="headers\IConfig.h" />
     <ClInclude Include="headers\ConfigFailure.h" />
     <ClInclude Include="headers\ConfigSuccess.h" />
-    <ClInclude Include="headers\IInputProvider.h" />
     <ClInclude Include="headers\InputProviderForMenu.h" />
     <ClInclude Include="headers\InputProviderForPlayerOne.h" />
     <ClInclude Include="headers\InputProviderForPlayerTwo.h" />
-    <ClInclude Include="headers\IObstacle.h" />
+    <ClInclude Include="headers\Interfaces\IConfig.h" />
+    <ClInclude Include="headers\Interfaces\IDrawable.h" />
+    <ClInclude Include="headers\Interfaces\IGame.h" />
+    <ClInclude Include="headers\Interfaces\IInputProvider.h" />
+    <ClInclude Include="headers\Interfaces\IMovable.h" />
+    <ClInclude Include="headers\Interfaces\IMoveBeh.h" />
+    <ClInclude Include="headers\Interfaces\IObstacle.h" />
+    <ClInclude Include="headers\Interfaces\ITickUpdatable.h" />
     <ClInclude Include="headers\Iron.h" />
     <ClInclude Include="headers\Map.h" />
-    <ClInclude Include="headers\ITickUpdatable.h" />
     <ClInclude Include="headers\Menu.h" />
     <ClInclude Include="headers\MouseButton.h" />
-    <ClInclude Include="headers\IMoveBeh.h" />
     <ClInclude Include="headers\MoveLikeAIBeh.h" />
     <ClInclude Include="headers\MoveLikeBulletBeh.h" />
     <ClInclude Include="headers\MoveLikeTankBeh.h" />
